Move button in CollectionDisplay (which is renamed from Collection for clarity)
