+++ client-v2/src/components/FrontsEdit/CollectionOverview.tsx
@@ -9,7 +9,7 @@ import { selectHasUnpublishedChanges } from 'selectors/frontsSelectors';
 import { openCollectionsAndFetchTheirArticles } from 'actions/Collections';
 
 import { Collection, CardSets } from 'shared/types/Collection';
-import { createCollectionId } from 'shared/components/Collection';
+import { createCollectionId } from 'shared/components/CollectionDisplay';
 import ButtonDefault from 'shared/components/input/ButtonCircular';
 import {
   createSelectCollection,
